# src/main/java/kz/halykacademy/bookstore/dto/BookNameDTO.java
package kz.halykacademy.bookstore.dto;

import lombok.*;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class BookNameDTO {
    private Long bookId;
    private String title;

}