# src/main/java/kz/halykacademy/bookstore/dto/SaveAuthorDTO.java
package kz.halykacademy.bookstore.dto;

import lombok.*;

import java.time.LocalDate;
import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class SaveAuthorDTO {

    private String fullName;
    private LocalDate dateOfBirth;


}