# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: /
...
paths:
  /authors/addAuthor:
    post:
      tags:
        - author
      summary: addAuthor
      operationId: addAuthorUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: author
          description: author
          required: true
          schema:
            $ref: '#/definitions/SaveAuthorDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AuthorDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /authors/allAuthors:
    get:
      tags:
        - author
      summary: findAll
      operationId: findAllUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AuthorDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /authors/deleteAuthor/{id}:
    delete:
      tags:
        - author
      summary: deleteAuthor
      operationId: deleteAuthorUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /authors/findByName/{name}:
    get:
      tags:
        - author
      summary: findByName
      operationId: findByNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AuthorDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /authors/getByGenres/{genre}:
    get:
      tags:
        - author
      summary: findAuthorsByGenreList
      operationId: findAuthorsByGenreListUsingGET
      produces:
        - '*/*'
      parameters:
        - name: genre
          in: path
          description: genre
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AuthorGenreListDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /authors/getById/{id}:
    get:
      tags:
        - author
      summary: getAuthorById
      operationId: getAuthorByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AuthorDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /authors/updateAuthor:
    put:
      tags:
        - author
      summary: updateAuthor
      operationId: updateAuthorUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: authorDTO
          description: authorDTO
          required: true
          schema:
            $ref: '#/definitions/SaveAuthorDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AuthorDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /books/addBook:
    post:
      tags:
        - book
      summary: addBook
      operationId: addBookUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: book
          description: book
          required: true
          schema:
            $ref: '#/definitions/SaveBookDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/BookDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /books/allBooks:
    get:
      tags:
        - book
      summary: getAllBooks
      operationId: getAllBooksUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/BookDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /books/delete/{id}:
    delete:
      tags:
        - book
      summary: deleteBook
      operationId: deleteBookUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /books/findByTitle/{title}:
    get:
      tags:
        - book
      summary: findByTitle
      operationId: findByTitleUsingGET
      produces:
        - '*/*'
      parameters:
        - name: title
          in: path
          description: title
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/BookDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /books/getByGenre/{genre}:
    get:
      tags:
        - book
      summary: findBookByGenre
      operationId: findBookByGenreUsingGET
      produces:
        - '*/*'
      parameters:
        - name: genre
          in: path
          description: genre
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/BookGenreDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /books/getById/{id}:
    get:
      tags:
        - book
      summary: getBookById
      operationId: getBookByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/BookDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /books/updateBook:
    put:
      tags:
        - book
      summary: updateBook
      operationId: updateBookUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: bookDTO
          description: bookDTO
          required: true
          schema:
            $ref: '#/definitions/SaveBookDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/BookDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /genres/addGenre:
    post:
      tags:
        - genre
      summary: addGenre
      operationId: addGenreUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: genre
          description: genre
          required: true
          schema:
            $ref: '#/definitions/GenreDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /genres/allGenres:
    get:
      tags:
        - genre
      summary: getAllGenres
      operationId: getAllGenresUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/GenreDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /genres/delete/{id}:
    delete:
      tags:
        - genre
      summary: deleteGenre
      operationId: deleteGenreUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /genres/getById/{id}:
    get:
      tags:
        - genre
      summary: getGenreById
      operationId: getGenreByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /genres/getByName/{name}:
    get:
      tags:
        - genre
      summary: findByName
      operationId: findByNameUsingGET_1
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /genres/updateGenre:
    put:
      tags:
        - genre
      summary: updateBook
      operationId: updateBookUsingPUT_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: genreDTO
          description: genreDTO
          required: true
          schema:
            $ref: '#/definitions/GenreDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenreDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /publishers/findByName/{name}:
    get:
      tags:
        - publisher
      summary: findByName
      operationId: findByNameUsingGET_2
      produces:
        - '*/*'
      parameters:
        - name: name
          in: path
          description: name
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/PublisherDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /publishers/updatePublisher:
    put:
      tags:
        - publisher
      summary: updateBook
      operationId: updateBookUsingPUT_2
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: publisherDTO
          description: publisherDTO
          required: true
          schema:
            $ref: '#/definitions/PublisherDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PublisherDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /users/addUser:
    post:
      tags:
        - user
      summary: addBook
      operationId: addBookUsingPOST_1
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userDTO
          description: userDTO
          required: true
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /users/delete/{id}:
    delete:
      tags:
        - user
      summary: deleteBook
      operationId: deleteBookUsingDELETE_1
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
...
definitions:
  AuthorDTO:
    type: object
    properties:
      books:
        type: array
        items:
          $ref: '#/definitions/BookNameDTO'
      dateOfBirth:
        type: string
        format: date
      fullName:
        type: string
      genreList:
        type: array
        items:
          type: string
      id:
        type: integer
        format: int64
    title: AuthorDTO
  AuthorGenreListDTO:
    type: object
    properties:
      fullName:
        type: string
      genreList:
        type: array
        items:
          type: string
    title: AuthorGenreListDTO
  AuthorNameDTO:
    type: object
    properties:
      fullName:
        type: string
      id:
        type: integer
        format: int64
    title: AuthorNameDTO
  BookDTO:
    type: object
    properties:
      authors:
        type: array
        items:
          $ref: '#/definitions/AuthorNameDTO'
      bookId:
        type: integer
        format: int64
      deleted:
        type: boolean
      genres:
        type: array
        items:
          $ref: '#/definitions/GenreNameDTO'
      pageCount:
        type: integer
        format: int32
      price:
        type: integer
        format: int32
      publisher:
        type: string
      releaseYear:
        type: string
        format: date
      title:
        type: string
    title: BookDTO
  BookGenreDTO:
    type: object
    properties:
      bookId:
        type: integer
        format: int64
      genres:
        type: array
        items:
          $ref: '#/definitions/GenreNameDTO'
      title:
        type: string
    title: BookGenreDTO
  BookNameDTO:
    type: object
    properties:
      bookId:
        type: integer
        format: int64
      title:
        type: string
    title: BookNameDTO
  GenreDTO:
    type: object
    properties:
      genreName:
        type: string
      id:
        type: integer
        format: int64
    title: GenreDTO
  GenreNameDTO:
    type: object
    properties:
      genreName:
        type: string
    title: GenreNameDTO
...
  PublisherDTO:
    type: object
    properties:
      name:
        type: string
      publisherId:
        type: integer
        format: int64
    title: PublisherDTO
  SaveAuthorDTO:
    type: object
    properties:
      dateOfBirth:
        type: string
        format: date
      fullName:
        type: string
      id:
        type: integer
        format: int64
    title: SaveAuthorDTO
  SaveBookDTO:
    type: object
    properties:
      authorList:
        type: array
        items:
          type: integer
          format: int64
      bookId:
        type: integer
        format: int64
      deleted:
        type: boolean
      genreList:
        type: array
        items:
          type: integer
          format: int64
      pageCount:
        type: integer
        format: int32
      price:
        type: integer
        format: int32
      publisherId:
        type: integer
        format: int64
      releaseYear:
        type: string
        format: date
      title:
        type: string
    title: SaveBookDTO
...
  UserDTO:
    type: object
    properties:
      blocked:
        type: boolean
      id:
        type: integer
        format: int64
      login:
        type: string
      password:
        type: string
      role:
        type: string
        enum:
          - USER
          - ADMIN
    title: UserDTO
...